delete model.load_weights function
this feature -> model.load_weights('model-best.h5')

By applying this feature, the analysis was limited only to stocks with the same number of rows, and other stocks with a different number of rows were not analyzed.  Also, I backtested by changing the ratio_cut value to 1, and as a result, it showed a better rate of return than 2.